Update bower package name to 'moment' for consistency
moment was registered with bower (via `bower register`) as both `moment` and `momentjs`. bower can't tell that these are the same packages. Consider a project that depends on `momentjs` and `moment-range`:

    my-application
    ├── momentjs (2.6.0)
    ├─┬ moment-range (2.6.0)
      └── moment (2.6.0)

Both `moment` and `momentjs` will be downloaded.

Considering that moment's own bower.json lists the package's name as `moment` [0], the NPM package is called `moment`, and important plugins depend on `moment` [1][2], update the docs to recommend `moment` when installing with bower.

[0]: https://github.com/moment/moment/blob/develop/bower.json#L2
[1]: https://github.com/gf3/moment-range/issues/38
[2]: https://github.com/icambron/twix.js/pull/38