Updated examples to of add() add(number, period)
> Deprecation warning: moment().add(period, number) is deprecated. Please use moment().add(number, period). 

This deprecation warning is useful, but annoying if you go directly off of the docs. Ideally, the docs should show the non-deprecated version unless specifically talking about older versions. I've changed the docs in this subsection. Let me know if you'd like me to do it for the rest of the doc, too; wanted to start small in case there is a scheduled release planned to fix this already.