Test documentation is runnable with markdown-doctest (#297)
markdown-doctest runs all of the javascript examples in the
documentation to ensure that they at least run without error. The idea
is to mimic the experience a user might have when copying and pasting
documentation.

I have add markdown-doctest to the deploy script so that the
documentation will only be deployed if there are no failures. This
should ensure that no future typos are introduced and if any breaking
changes are made, all relevant pieces of the documentation should be
easy to update.