Correct rounding to truncating (#469)
According to the tests and the implementation, Moment.js does not round
but truncates the result, e.g. 1.9 results in 1 instead of 2. The
expression "round towards zero" might leave the impression that the
behaviour changed in 2.0.0 only for values n + 1/2 where n is an
integer.

The example is changed to illustrate the design choice more clearly.

This commit also includes additional formatting fixes.