Update Customise parent locale
This adds the description of defining a locale with a parent that hasn't itself been defined or loaded yet which has been available since 2.16.0. This adds the additional behavior when creating a moment of lazy loading the parent locale if it isn't already loaded, or defaulting to the global locale if the parent doesn't exist. See PR (https://github.com/moment/moment/pull/4310) As the next version number isn't determined yet, this should be merged only with the next release that includes the above PR.