Structure proposal for SECURITY.md
Related to #580 

Having a Security Policy can help to set a private way to vulnerability reports, keep in mind that Moment is a dependency of many projects, and sometimes making a vulnerability public since the begin is not the best way to handle it, attackers will be able to act before the community will be able to handle and fix it. 

If you folks care about security, setting a Security Policy is a good way to show this to the community.