markdown formatting with usages of "javascript"
usage of ```javascript``` in markdown code snippets works in github previews but breaks the formatting used in the docs themselves on https://momentjs.com/docs. Other code snippets in the docs do not use this syntax and therefore for consistency and to fix the broken formatting on the docs site, this snippet should remove the above formatting.